+++ src/main/java/bdv/tools/brightness/colorscheme/AbstractColorScheme.java
@@ -33,7 +33,7 @@ import net.imglib2.type.numeric.ARGBType;
  * Shared color-stop storage and RGB/RGBA mechanics for
  * {@link DiscreteColorScheme} and {@link ContinuousColorScheme}: the two only
  * differ in how a palette value resolves to a stop (or a blend of two) --
- * see {@link #colorAt(float)} -- and in {@link ColorScheme#getPaletteRangeLength()}.
+ * see {@link #colorAt(double)} -- and in {@link ColorScheme#getPaletteRangeLength()}.
  * Package-private: an implementation detail, not part of the public API in
  * {@link ColorScheme}.
  */
@@ -61,13 +61,13 @@ abstract class AbstractColorScheme implements ColorScheme
 	}
 
 	@Override
-	public final int getRGBA( final float paletteValue )
+	public final int getRGBA( final double paletteValue )
 	{
 		return colorAt( paletteValue );
 	}
 
 	@Override
-	public final int getRGB( final float paletteValue )
+	public final int getRGB( final double paletteValue )
 	{
 		return colorAt( paletteValue ) | 0xff000000;
 	}
@@ -79,10 +79,10 @@ abstract class AbstractColorScheme implements ColorScheme
 	 * (or half-open equivalent) resolves to its nearest edge stop rather than
 	 * throwing or reading out of bounds.
 	 */
-	abstract int colorAt( float paletteValue );
+	abstract int colorAt( double paletteValue );
 
 	/** Linearly interpolates each channel independently between two packed-ARGB stops, {@code t} in {@code [0, 1]}. */
-	static int interpolateColor( final int fromARGB, final int toARGB, final float t )
+	static int interpolateColor( final int fromARGB, final int toARGB, final double t )
 	{
 		final int r = interpolateChannel( ARGBType.red( fromARGB ), ARGBType.red( toARGB ), t );
 		final int g = interpolateChannel( ARGBType.green( fromARGB ), ARGBType.green( toARGB ), t );
@@ -91,8 +91,8 @@ abstract class AbstractColorScheme implements ColorScheme
 		return ARGBType.rgba( r, g, b, a );
 	}
 
-	private static int interpolateChannel( final int from, final int to, final float t )
+	private static int interpolateChannel( final int from, final int to, final double t )
 	{
-		return Math.round( from + t * ( to - from ) );
+		return ( int ) Math.round( from + t * ( to - from ) );
 	}
 }
