+++ src/main/java/bdv/tools/brightness/presetfunc/SigmoidPresetFunc.java
@@ -35,7 +35,7 @@ public class SigmoidPresetFunc extends AbstractPresetFunc
 {
 	private static final double K = 10.0;
 
-	public SigmoidPresetFunc( final float min, final float max, final int paletteRangeLength )
+	public SigmoidPresetFunc( final double min, final double max, final int paletteRangeLength )
 	{
 		super( min, max, paletteRangeLength );
 	}
@@ -47,7 +47,7 @@ public class SigmoidPresetFunc extends AbstractPresetFunc
 	}
 
 	@Override
-	public SigmoidPresetFunc withRange( final float min, final float max )
+	public SigmoidPresetFunc withRange( final double min, final double max )
 	{
 		return new SigmoidPresetFunc( min, max, getPaletteRangeLength() );
 	}
